use ModularMain in the example modules (#4293)
All examples compile, and if you edit the JSON files to use absolute paths, all the demos still work (and if you don't use absolute paths, they fail on the `main` branch, too!). 

The intention here is that I've made a `ModularMain()` function (see https://github.com/viamrobotics/rdk/pull/4256) to avoid copying and pasting so much boilerplate when running modules, and we should encourage users to use it instead of copying and pasting boilerplate, too.