[RSDK-8494] Validate PWM-setting requests (#4309)
The Jira ticket was specifically for the pinctrl module, but I think it should be generalized to all boards. So, here we go.

Thanks to John for the suggestion of calling "slightly over 1.0" a non-error value! If you want a different threshold for "slightly over," I'm happy to change it.

The rpi4 doesn't validate that the frequency is positive because apparently setting it to 0 already sets it to the default of 800 Hz, and it's a uint so can't be negative.

I haven't tested on actual hardware, but the new unit tests pass. I'm willing to test on hardware if you think it's important.