use StoppableWorkers in vectornav/imu (#4318)
This all compiles, but I don't know how to test it out on real hardware. I'm willing to do that if you think it's important, but my hope is the change is simple enough that we can skip it.

I don't think this PR actually fixes any bugs or race conditions: the old code looked right to me. but I also think the new version is simpler to reason about.

Github suggests that Dan G should review this; if you disagree, feel free to remove yourself and add someone else.