[RSDK-10646] Allow uploading a module via CLI if the windows build's meta.json doesn't have a ".exe" in the entrypoint (#4969)
I wanted to write a test for this, but I also wanted not to add any binary files (including .tar.gz files) into the repo, and couldn't figure out how to square that circle. This function is not currently tested, so I'm not decreasing the test coverage very much.

Instead, I built the CLI both with and without this change, and tried uploading a Windows module: this version succeeded, whereas the `main` branch failed with an `Error validating module: the archive does not contain a file at the desired entrypoint`.