more efficient implementation of BasicOctree.PointsCollidingWith() (#5281)
I got annoyed that the sanding tests kept using so much memory that my computer would become sluggish and sometimes my web browser would crash from OOM. So, I profiled the tests, and found that the literal majority(!) of memory was used in `BasicOctree.PointsCollidingWith()`. It uses so much memory because every time it recurses from a parent node to a child node, the child starts building the list of points anew, and then the parent copies everything the child returns into its own list.

(This is a problem because Go uses mark-and-sweep garbage collection: all copies of temporary variables remain allocated until the garbage collector finishes deciding what data is currently live, so large temporary variables can add up if allocated quickly.)

With this PR, we create just one list of points, and pass it around as an accumulator: recursing into a child node passes the pointer to the list, and then the child adds any relevant data to that list directly. Then at the very end, we return this one list of points, without ever making a second copy.

Tried in the sanding ICP tests: instead of allocating 167 gigabytes(!) of memory, this function now allocates "only" 26 gigabytes. I still don't like how memory-intensive those tests are, but this is an easy improvement. 

## Old memory profile of ICP tests: 
<img width="646" height="772" alt="Screenshot_2025-09-12_14-25-34" src="https://github.com/user-attachments/assets/e0b450b2-c2ba-4590-b029-6882900ac16f" />

## New memory profile:
<img width="784" height="646" alt="Screenshot_2025-09-12_14-25-49" src="https://github.com/user-attachments/assets/4c0f6f76-966a-4d14-a5d4-35653903c014" />